requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
Authlib==1.3.2
backoff==2.2.1
botocore==1.35.27
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
distlib==0.3.8
Django==5.1.1
filelock==3.16.1
frozenlist==1.4.1
gql==3.5.0
graphql-core==3.2.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jmespath==1.0.1
multidict==6.1.0
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
urllib3==2.2.3
virtualenv==20.26.5
websockets==11.0.3
yarl==1.12.1
